+++ src/routes/Login/index.js
@@ -3,7 +3,7 @@ import { dynamicWrapper, createRoute } from '@/utils/core';
 const routesConfig = (app) => ({
   path: '/sign/login',
   title: 'Login',
-  component: dynamicWrapper(app, [import('./model')], () => import('./components')),
+  component: dynamicWrapper(app, [import('./model')], () => import('./components'))
 });
 
 export default (app) => createRoute(app, routesConfig);
