+++ src/routes/Widgets/Gallery/components/index.js
