+++ src/routes/Widgets/Result/components/index.js
