+++ src/routes/Widgets/Result/index.js
