+++ src/routes/index.js
@@ -2,12 +2,12 @@ import { createRoutes } from '@/utils/core';
 import BasicLayout from '@/layouts/BasicLayout';
 import UserLayout from '@/layouts/UserLayout';
 import NotFound from './Pages/404';
-import ScreenLock from './Pages/ScreenLock';
 import Page500 from './Pages/500';
-import Coming from './Pages/Coming';
-import Gallery from './Pages/Gallery';
-import Result from './Pages/Result';
-import LevelRoute from './Pages/LevelRoute';
+import ScreenLock from './Widgets/ScreenLock';
+import Coming from './Widgets/Coming';
+import Gallery from './Widgets/Gallery';
+import Result from './Widgets/Result';
+import LevelRoute from './Widgets/LevelRoute';
 import Login from './Login';
 import Register from './Register';
 import Dashboard from './Dashboard';
